rpc-test: move wallet rpc tests to wallet-rpc-tests file.
